Removing Conditional assignments on buffers
Conditional assignment is unavoidably useful in ASIC design to infer clock gating control.
But, in FPGA, the clock fabric is fixed. The conditional assignment merely incurs additional logic area because the buffer flops now need to feedback the "steady state" value to be retained when the assignment conditional is false.