LSU Work, HCU shelfed
- Completing LSU with store functionality, unit test updated
- Stripping out hcu from ID. HCU will require significant rework. Will
need to re-architect how hazards are handled in kronos. The plan is to
entirely do all of the hazard control in the ID stage, such that EX
stage will always recieve valid operands. This exploration started from
early exploration of an implementation of the kronos core in an
ice40up5k. The EX is truly the critical path. HCU unit test currently
failing