HCU
- Brand new Hazard Control Unit, integrated into the Decode stage.
Comments in the kronos_hcu explain its operation. All hazards are
managed at the decode stage, and only clean instructions make it to the
EX stage. Thus, the EX stage no longer needs to watch out for hazards.
The decode stage stalls on hazard.
- Dedicated hcu unit test added, and previous hcu test which instanced
that kronos core move to core_basic
- Write Back updated to Zero cycle as it used to be, since the logic
doesn't add to the critical path of the EX stage.